requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
altair==4.2.2 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
blinker==1.6.2 ; python_version >= '3.7'
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
db-dtypes==1.1.1
decorator==5.1.1 ; python_version >= '3.5'
entrypoints==0.4 ; python_version >= '3.6'
frozenlist==1.3.3 ; python_version >= '3.7'
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.31 ; python_version >= '3.7'
google-api-core[grpc]==2.11.0 ; python_version >= '3.7'
google-auth==2.19.0 ; python_version >= '3.6'
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.5.0 ; python_version >= '3.7'
googleapis-common-protos==1.59.0 ; python_version >= '3.7'
grpcio==1.54.2 ; python_version >= '3.7'
grpcio-status==1.48.2
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.6.0 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.17.3 ; python_version >= '3.7'
markdown-it-py==2.2.0 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.24.3 ; python_version >= '3.8'
openai==0.27.7
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.1 ; python_version >= '3.8'
pillow==9.5.0 ; python_version >= '3.7'
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==3.20.3 ; python_version >= '3.7'
pyarrow==12.0.0 ; python_version >= '3.7'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydeck==0.8.1b0 ; python_version >= '3.7'
pygments==2.15.1 ; python_version >= '3.7'
pympler==1.0.1 ; python_version >= '3.6'
pyrsistent==0.19.3 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
requests==2.31.0 ; python_version >= '3.7'
rich==13.3.5 ; python_full_version >= '3.7.0'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0 ; python_version >= '3.6'
streamlit==1.22.0
tenacity==8.2.2 ; python_version >= '3.6'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0 ; python_version >= '3.5'
tornado==6.3.2 ; python_version >= '3.8'
tqdm==4.65.0 ; python_version >= '3.7'
typing-extensions==4.6.2 ; python_version >= '3.7'
tzdata==2023.3 ; python_version >= '2'
tzlocal==5.0.1 ; python_version >= '3.7'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
validators==0.20.0 ; python_version >= '3.4'
yarl==1.9.2 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
